Add scripts for KVM virtual machine automation
This commit adds three scripts for automating the creation and configuration of virtual machines in KVM. The scripts included are:
- crearMaquina.sh: Creates a virtual machine
- default-net.sh: Creates a default virtual network
- static-bridge.sh: Creates a virtual network with a static bridge

The scripts can be used to easily create virtual machines with different operating systems. The user can provide the necessary parameters such as the name, RAM, hard disk size, operating system, number of CPUs, and network to connect to.

The scripts also provide options to view and modify the configuration of the virtual networks.

Author: Adrián Cabezuelo Expósito